+++ api/src/certification/evaluation/domain/usecases/index.js
@@ -8,8 +8,6 @@ import * as languageService from '../../../../shared/domain/services/language-se
 import * as placementProfileService from '../../../../shared/domain/services/placement-profile-service.js';
 import { injectDependencies } from '../../../../shared/infrastructure/utils/dependency-injection.js';
 import { importNamedExportsFromDirectory } from '../../../../shared/infrastructure/utils/import-named-exports-from-directory.js';
-import * as enrolmentCertificationCandidateRepository from '../../../enrolment/infrastructure/repositories/certification-candidate-repository.js';
-import * as userRepository from '../../../enrolment/infrastructure/repositories/user-repository.js';
 import * as flashAlgorithmService from '../../../flash-certification/domain/services/algorithm-methods/flash.js';
 import {
   answerRepository,
@@ -24,8 +22,10 @@ import {
   sharedCompetenceMarkRepository,
 } from '../../../session-management/infrastructure/repositories/index.js';
 import * as certificationBadgesService from '../../../shared/domain/services/certification-badges-service.js';
+import * as sharedCertificationCandidateRepository from '../../../shared/infrastructure/repositories/certification-candidate-repository.js';
 import * as certificationCenterRepository from '../../../shared/infrastructure/repositories/certification-center-repository.js';
 import * as certificationCourseRepository from '../../../shared/infrastructure/repositories/certification-course-repository.js';
+import * as userRepository from '../../../shared/infrastructure/repositories/user-repository.js';
 import * as certificationCandidateRepository from '../../infrastructure/repositories/certification-candidate-repository.js';
 import * as certificationCompanionAlertRepository from '../../infrastructure/repositories/certification-companion-alert-repository.js';
 /**
@@ -35,7 +35,7 @@ import * as certificationCompanionAlertRepository from '../../infrastructure/rep
 const dependencies = {
   ...sessionRepositories,
   assessmentRepository,
-  enrolmentCertificationCandidateRepository,
+  sharedCertificationCandidateRepository,
   verifyCertificateCodeService,
   assessmentResultRepository,
   answerRepository,
