+++ api/src/certification/shared/infrastructure/repositories/user-repository.js
@@ -2,7 +2,7 @@ import _ from 'lodash';
 
 import { knex } from '../../../../../db/knex-database-connection.js';
 import { DomainTransaction } from '../../../../shared/domain/DomainTransaction.js';
-import { User } from '../../domain/models/User.js';
+import { User } from '../../../enrolment/domain/models/User.js';
 
 export async function get({ id }) {
   const knexConn = DomainTransaction.getConnection();
